requirements.txt
ansi2html==1.8.0
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
Flask==2.2.5
fonttools==4.43.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipydatawidgets==4.3.5
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
libigl==2.4.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nbformat==5.7.0
nest-asyncio==1.5.8
numpy==1.26.1
open3d==0.17.0
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pyglet==1.5.28
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pythreejs==2.4.2
pywin32==306
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rpds-py==0.10.6
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tornado==6.3.3
traitlets==5.11.2
traittypes==0.2.1
transforms3d==0.4.1
trimesh==4.0.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==2.2.3
widgetsnbextension==4.0.9
zipp==3.17.0
